transpile flow in node_modules and remove v3 as well as its identical to v4
